Lazy load images
Plus bunch of other improvements and general code clean-up.

Images are loaded lazily by a simple hack that stores all image sources
in a data-src= attribute that is then converted to an actual src=
attribute once the image is displayed.

Also, arrows are now base64 encoded strings so no need to copy any SVG
files. The navigation also has a counter in the middle to indicate
number of photos.